requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==3.0.5
charset-normalizer==3.4.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.1
daphne==3.0.2
Django==5.1.2
django-4-jet==1.0.9
django-cors-headers==4.5.0
django-environ==0.11.2
django-filter==24.3
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.1
openapi-codec==1.3.2
packaging==24.1
pillow==11.0.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
service-identity==24.1.0
setuptools==75.1.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.7.0
zope.interface==7.1.0
